+++ include/jian/pdb/Manipulate.h
@@ -7,7 +7,7 @@
 namespace jian {
 namespace pdb {
     
-inline Vector3f normal_vector(const Residue &res) {
+inline Vector3d normal_vector(const Residue &res) {
     if (std::set<std::string>{"A", "G", "DA", "DG"}.count(res.name)) {
         auto atom_list = res[std::vector<std::string>{"C2", "C4", "C6"}];
         return geom::normal_vector(atom_list[0], atom_list[1], atom_list[2]);
@@ -17,11 +17,11 @@ inline Vector3f normal_vector(const Residue &res) {
     }
 }
 
-inline Vector3f pivot(const Residue &res) {
+inline Vector3d pivot(const Residue &res) {
     if (std::set<std::string>{"A", "G", "DA", "DG"}.count(res.name)) {
-        return res["N9"].pos<Vector3f>();
+        return res["N9"].pos<Vector3d>();
     } else if (std::set<std::string>{"U", "C", "DT", "DC"}.count(res.name)) {
-        return res["N1"].pos<Vector3f>();
+        return res["N1"].pos<Vector3d>();
     }
 }
